finalization workflow logging update

+++ db/migrate/20160127165855_drop_admin_user.rb
@@ -1,5 +1,9 @@
 class DropAdminUser < ActiveRecord::Migration
-  def change 
-     drop_table :admin_users
+  def up
+     drop_table :admin_users if ActiveRecord::Base.connection.table_exists? 'admin_users'
+  end
+
+  def down
+     #not reversable
   end
 end
