+++ db/migrate/20160202162321_drop_dvd_delivery.rb
@@ -1,15 +1,14 @@
 class DropDvdDelivery < ActiveRecord::Migration
   def up
-     remove_foreign_key :orders, :dvd_delivery_location
-     remove_column :orders, :dvd_delivery_location_id
-     drop_table :dvd_delivery_locations
+     if ActiveRecord::Base.connection.table_exists? 'dvd_delivery_locations'
+        remove_foreign_key :orders, :dvd_delivery_location
+        remove_column :orders, :dvd_delivery_location_id
+        drop_table :dvd_delivery_locations
+     end
   end
 
   def down
-     create_table :dvd_delivery_locations do |t|
-        t.string :name
-        t.string :email_desc
-     end
-     add_column :orders, :dvd_delivery_location_id, :integer
+     # not reversable
   end
+
 end
