+++ db/migrate/20160208213347_remove_archive.rb
@@ -1,17 +1,13 @@
 class RemoveArchive < ActiveRecord::Migration
    def up
-      remove_foreign_key :units, :archive
-      remove_column :units, :archive_id
-      drop_table :archives
+      if ActiveRecord::Base.connection.table_exists? 'archives'
+         remove_foreign_key :units, :archive
+         remove_column :units, :archive_id
+         drop_table :archives
+      end
    end
 
    def down
-      create_table :archives do |t|
-         t.string :name
-         t.string :directory
-         t.string :description
-         t.integer :units_count
-      end
-      add_column :units, :archive_id, :integer
+      # not reversable
    end
 end
