+++ db/migrate/20160209204329_remove_unit_import_source.rb
@@ -1,14 +1,9 @@
 class RemoveUnitImportSource < ActiveRecord::Migration
    def up
-      drop_table :unit_import_sources
+      drop_table :unit_import_sources if ActiveRecord::Base.connection.table_exists? 'unit_import_sources'
    end
 
    def down
-      create_table :unit_import_sources do |t|
-         t.integer :unit_id
-         t.string :standard
-         t.string :version
-         t.text :source
-      end
+      # not reversable
    end
 end
