+++ db/migrate/20160209205933_remove_session_table.rb
@@ -1,5 +1,9 @@
 class RemoveSessionTable < ActiveRecord::Migration
-  def change
-     drop_table :sessions
+  def up
+     drop_table :sessions if ActiveRecord::Base.connection.table_exists? 'sessions'
+  end
+
+  def down
+     # not reversable
   end
 end
