+++ db/migrate/20160209212245_drop_delivery_methods.rb
@@ -1,7 +1,12 @@
 class DropDeliveryMethods < ActiveRecord::Migration
-   def change
+   def up
+      if ActiveRecord::Base.connection.table_exists? 'delivery_methods'
+         drop_table :delivery_methods_orders
+         drop_table :delivery_methods
+      end
+   end
 
-      drop_table :delivery_methods_orders
-      drop_table :delivery_methods
+   def down
+      # not reversable
    end
 end
